Refactor: simplify configuration code.
All the shared code, including for case insensivity
and related finding almost-duplicates (grouos and projects
configurations where group or project names differ only
in case) is now in 'core'.

The common, groups and projects-related methods are now
groupped into a separate class each. They inherit from
themselves this way: common -> groups -> projects,
because this is how the configurations depend in each
other.
