+++ gitlabform/configuration/core.py
@@ -20,6 +20,11 @@ from ruamel.yaml.comments import CommentedMap
 
 
 class ConfigurationCore(ABC):
+    """
+    The functionality shared by various types of config (common, groups, projects)
+    is implemented here.
+    """
+
     def __init__(self, config_path=None, config_string=None):
 
         if config_path and config_string:
@@ -56,8 +61,10 @@ class ConfigurationCore(ABC):
                         exit_code=EXIT_INVALID_INPUT,
                     )
 
-                # we are NOT checking for the existence of non-empty 'projects_and_groups' key here
-                # as it would break using GitLabForm as a library
+            self._find_almost_duplicates()
+
+            # we are NOT checking for the existence of non-empty 'projects_and_groups' key here
+            # as it would break using GitLabForm as a library
 
         except (FileNotFoundError, OSError):
             raise ConfigFileNotFoundException(config_path)
@@ -149,33 +156,8 @@ class ConfigurationCore(ABC):
         else:
             return current
 
-    def get_group_config(self, group) -> dict:
-        pass
-
-    def get_project_config(self, group_and_project) -> dict:
-        pass
-
-    def get_common_config(self) -> dict:
-        """
-        :return: literal common configuration or empty dict if not defined
-        """
-        common_config = self.get("projects_and_groups|*", {})
-        if common_config:
-            section_name = "*"
-            self.validate_break_inheritance_flag(common_config, section_name)
-        return common_config
-
-    def is_project_skipped(self, project) -> bool:
-        pass
-
-    def is_group_skipped(self, group) -> bool:
-        pass
-
-    def is_skipped(self, an_array: list, item: str) -> bool:
-        pass
-
     @staticmethod
-    def validate_break_inheritance_flag(
+    def _validate_break_inheritance_flag(
         config: dict, section_name: str, parent_key: str = ""
     ):
         for key, value in config.items():
@@ -189,12 +171,12 @@ class ConfigurationCore(ABC):
                     exit_code=EXIT_INVALID_INPUT,
                 )
             elif type(value) is CommentedMap:
-                ConfigurationCore.validate_break_inheritance_flag(
+                ConfigurationCore._validate_break_inheritance_flag(
                     value, section_name, key
                 )
 
     @staticmethod
-    def merge_configs(more_general_config, more_specific_config) -> dict:
+    def _merge_configs(more_general_config, more_specific_config) -> dict:
         """
         :return: merge more general config with more specific configs.
                  More specific config values take precedence over more general ones.
@@ -233,6 +215,95 @@ class ConfigurationCore(ABC):
         break_inheritance(more_specific_config)
 
         return dict(merged_dict)
+
+    @staticmethod
+    def _get_case_insensitively(a_dict: dict, a_key: str):
+        for dict_key in a_dict.keys():
+            if dict_key.lower() == a_key.lower():
+                return a_dict[dict_key]
+        return {}
+
+    @staticmethod
+    def _is_skipped_case_insensitively(an_array: list, item: str) -> bool:
+        """
+        :return: if item is defined in the list to be skipped
+        """
+        item = item.lower()
+
+        for list_element in an_array:
+            list_element = list_element.lower()
+
+            if list_element == item:
+                return True
+
+            if (
+                list_element.endswith("/*")
+                and item.startswith(list_element[:-2])
+                and len(item) >= len(list_element[:-2])
+            ):
+                return True
+
+        return False
+
+    def _find_almost_duplicates(self):
+
+        # in GitLab groups and projects names are de facto case insensitive:
+        # you can change the case of both name and path BUT you cannot create
+        # 2 groups which names differ only with case and the same thing for
+        # projects. therefore we cannot allow such entries in the config,
+        # as they would be ambiguous.
+
+        for path in [
+            "projects_and_groups",
+            "skip_groups",
+            "skip_projects",
+        ]:
+            if self.get(path, 0):
+                almost_duplicates = self._find_almost_duplicates_in(path)
+                if almost_duplicates:
+                    fatal(
+                        f"There are almost duplicates in the keys of {path} - they differ only in case.\n"
+                        f"They are: {', '.join(almost_duplicates)}\n"
+                        f"This is not allowed as we ignore the case for group and project names.",
+                        exit_code=EXIT_INVALID_INPUT,
+                    )
+
+    def _find_almost_duplicates_in(self, configuration_path):
+        """
+        Checks given configuration key and reads its keys - if it is a dict - or elements - if it is a list.
+        Looks for items that are almost the same - they differ only in the case.
+        :param configuration_path: configuration path, f.e. "group_settings"
+        :return: list of items that have almost duplicates,
+                 or an empty list if none are found
+        """
+
+        dict_or_list = self.get(configuration_path)
+        if isinstance(dict_or_list, dict):
+            items = dict_or_list.keys()
+        else:
+            items = dict_or_list
+
+        items_with_lowercase_names = [x.lower() for x in items]
+
+        # casting these to sets will deduplicate the one with lowercase names
+        # lowering its cardinality if there were elements in it
+        # that before lowering differed only in case
+        if len(set(items)) != len(set(items_with_lowercase_names)):
+
+            # we have some almost duplicates, let's find them
+            almost_duplicates = []
+            for first_item in items:
+                occurrences = 0
+                for second_item in items_with_lowercase_names:
+                    if first_item.lower() == second_item.lower():
+                        occurrences += 1
+                        if occurrences == 2:
+                            almost_duplicates.append(first_item)
+                            break
+            return almost_duplicates
+
+        else:
+            return []
 
 
 class ConfigFileNotFoundException(Exception):
