+++ tests/unit/test_non_empty_configs_provider.py
@@ -1,7 +1,7 @@
 import pytest
 
 from gitlabform import EXIT_INVALID_INPUT
-from gitlabform.configuration.projects_and_groups import ConfigurationProjectsAndGroups
+from gitlabform.configuration import Configuration
 from gitlabform.filter import NonEmptyConfigsProvider
 
 
@@ -12,7 +12,7 @@ def test_error_on_missing_key():
     """
 
     with pytest.raises(SystemExit) as e:
-        configuration = ConfigurationProjectsAndGroups(config_string=config_yaml)
+        configuration = Configuration(config_string=config_yaml)
         NonEmptyConfigsProvider(configuration, None, None)
     assert e.value.code == EXIT_INVALID_INPUT
 
@@ -24,6 +24,6 @@ def test_error_on_empty_key():
     """
 
     with pytest.raises(SystemExit) as e:
-        configuration = ConfigurationProjectsAndGroups(config_string=config_yaml)
+        configuration = Configuration(config_string=config_yaml)
         NonEmptyConfigsProvider(configuration, None, None)
     assert e.value.code == EXIT_INVALID_INPUT
